Refactor exception package structure and update imports for consistency
