+++ domain/model/src/main/java/co/com/pragma/model/exceptions/CustomException.java
@@ -1,4 +1,4 @@
-package co.com.pragma.model.solicitude.exceptions;
+package co.com.pragma.model.exceptions;
 
 import lombok.Getter;
