+++ domain/model/src/main/java/co/com/pragma/model/exceptions/ValueOutOfBoundsException.java
@@ -1,4 +1,4 @@
-package co.com.pragma.model.solicitude.exceptions;
+package co.com.pragma.model.exceptions;
 
 import co.com.pragma.model.constants.Errors;
