+++ domain/model/src/main/java/co/com/pragma/model/loantype/exceptions/LoanTypeNotFoundException.java
@@ -1,7 +1,7 @@
 package co.com.pragma.model.loantype.exceptions;
 
 import co.com.pragma.model.constants.Errors;
-import co.com.pragma.model.solicitude.exceptions.CustomException;
+import co.com.pragma.model.exceptions.CustomException;
 
 public class LoanTypeNotFoundException extends CustomException {
     public LoanTypeNotFoundException() {
