+++ domain/model/src/main/java/co/com/pragma/model/loantype/exceptions/LoanTypeValueErrorException.java
@@ -1,7 +1,7 @@
 package co.com.pragma.model.loantype.exceptions;
 
 import co.com.pragma.model.constants.Errors;
-import co.com.pragma.model.solicitude.exceptions.CustomException;
+import co.com.pragma.model.exceptions.CustomException;
 
 public class LoanTypeValueErrorException extends CustomException {
     public LoanTypeValueErrorException() {
