+++ domain/model/src/main/java/co/com/pragma/model/solicitude/exceptions/SolicitudeNullException.java
@@ -1,6 +1,7 @@
 package co.com.pragma.model.solicitude.exceptions;
 
 import co.com.pragma.model.constants.Errors;
+import co.com.pragma.model.exceptions.CustomException;
 
 public class SolicitudeNullException extends CustomException {
     public SolicitudeNullException() {
