--- domain/model/src/main/java/co/com/pragma/model/user/Role.java
@@ -1,14 +1,0 @@
-package co.com.pragma.model.user;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder(toBuilder = true)
-public class Role {
-    private Integer rolId;
-    private String name;
-    private String description;
-}
