--- domain/model/src/main/java/co/com/pragma/model/user/User.java
@@ -1,27 +1,0 @@
-package co.com.pragma.model.user;
-import lombok.Builder;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-import java.time.LocalDate;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder(toBuilder = true)
-public class User {
-    private Integer userId;
-    private String name;
-    private String lastName;
-    private String email;
-    private String idNumber;
-    private Role role;
-    private BigDecimal baseSalary;
-    private String phone;
-    private String address;
-    private LocalDate birthDate;
-}
