+++ domain/model/src/main/java/co/com/pragma/model/user/gateways/UserPort.java
@@ -1,8 +1,4 @@
 package co.com.pragma.model.user.gateways;
 
-import co.com.pragma.model.user.User;
-import reactor.core.publisher.Mono;
-
 public interface UserPort {
-    Mono<User> getUserByIdNumber(String idNumber);
 }
