+++ domain/usecase/src/test/java/co/com/pragma/usecase/solicitude/SolicitudeUseCaseTest.java
@@ -7,9 +7,9 @@ import co.com.pragma.model.loantype.exceptions.LoanTypeNotFoundException;
 import co.com.pragma.model.loantype.gateways.LoanTypeRepository;
 import co.com.pragma.model.logs.gateways.LoggerPort;
 import co.com.pragma.model.solicitude.Solicitude;
-import co.com.pragma.model.solicitude.exceptions.InvalidFieldException;
+import co.com.pragma.model.exceptions.InvalidFieldException;
 import co.com.pragma.model.solicitude.exceptions.SolicitudeNullException;
-import co.com.pragma.model.solicitude.exceptions.ValueOutOfBoundsException;
+import co.com.pragma.model.exceptions.ValueOutOfBoundsException;
 import co.com.pragma.model.solicitude.gateways.SolicitudeRepository;
 import co.com.pragma.model.state.State;
 import co.com.pragma.model.state.exceptions.StateNotFoundException;
