+++ domain/usecase/src/test/java/co/com/pragma/usecase/solicitude/utils/SolicitudeUtilsTest.java
@@ -4,9 +4,9 @@ import co.com.pragma.model.constants.DefaultValues;
 import co.com.pragma.model.loantype.LoanType;
 import co.com.pragma.model.loantype.exceptions.LoanTypeValueErrorException;
 import co.com.pragma.model.solicitude.Solicitude;
-import co.com.pragma.model.solicitude.exceptions.FieldBlankException;
-import co.com.pragma.model.solicitude.exceptions.FieldSizeOutOfBounds;
-import co.com.pragma.model.solicitude.exceptions.ValueOutOfBoundsException;
+import co.com.pragma.model.exceptions.FieldBlankException;
+import co.com.pragma.model.exceptions.FieldSizeOutOfBounds;
+import co.com.pragma.model.exceptions.ValueOutOfBoundsException;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import reactor.test.StepVerifier;
